style: update with reusable xss validation service
